### 2025-09-10 #8969 @editorialbot
[PRE REVIEW]: FSML – A Modern Fortran Statistics and Machine Learning Library
**Submitting author:** @sebastian-mutz (<a href="http://orcid.org/0000-0001-8180-6150">Sebastian G. Mutz</a>)
**Repository:** https://github.com/sebastian-mutz/fsml
**Branch with paper.md** (empty if default branch):
**Version:** v0.0.9
