### 2025-09-10 119fc41
fixed doi's and word count
