+++ paper/refs.bib
@@ -2,7 +2,8 @@
   title={AIFS-ECMWF's data-driven forecasting system},
   author={Lang, Simon and Alexe, Mihai and Chantry, Matthew and Dramsch, Jesper and Pinault, Florian and Raoult, Baudouin and Clare, Mariana CA and Lessig, Christian and Maier-Gerber, Michael and Magnusson, Linus and others},
   journal={arXiv preprint arXiv:2406.01465},
-  year={2024}
+  year={2024},
+  doi = {10.48550/arXiv.2406.01465}
 }
 
 @article{boateng:2023,
@@ -55,7 +56,7 @@ number = {9},
 pages = {100340},
 year = {2021},
 issn = {2666-3899},
-doi = {https://doi.org/10.1016/j.patter.2021.100340},
+doi = {10.1016/j.patter.2021.100340},
 url = {https://www.sciencedirect.com/science/article/pii/S2666389921001884},
 author = {Charlotte Freitag and Mike Berners-Lee and Kelly Widdicks and Bran Knowles and Gordon S. Blair and Adrian Friday},
 keywords = {ICT, carbon footprint, trends, AI, big data, data science, IoT, blockchain, policy, regulations},
@@ -70,7 +71,7 @@ number = {7},
 pages = {1613-1637},
 year = {2018},
 keywords = {ICON-A, atmospheric GCM, model description, model tuning},
-doi = {https://doi.org/10.1029/2017MS001242},
+doi = {10.1029/2017MS001242},
 url = {https://agupubs.onlinelibrary.wiley.com/doi/abs/10.1029/2017MS001242},
 eprint = {https://agupubs.onlinelibrary.wiley.com/doi/pdf/10.1029/2017MS001242},
 }
@@ -82,7 +83,7 @@ volume = {205},
 pages = {102609},
 year = {2021},
 issn = {0167-6423},
-doi = {https://doi.org/10.1016/j.scico.2021.102609},
+doi = {10.1016/j.scico.2021.102609},
 url = {https://www.sciencedirect.com/science/article/pii/S0167642321000022},
 author = {Rui Pereira and Marco Couto and Francisco Ribeiro and Rui Rua and Jácome Cunha and João Paulo Fernandes and João Saraiva},
 keywords = {Energy efficiency, Programming languages, Language benchmarking, Green software},
@@ -145,7 +146,7 @@ number = {1},
 pages = {14-26},
 year = {2009},
 issn = {0022-1694},
-doi = {https://doi.org/10.1016/j.jhydrol.2008.12.017},
+doi = {10.1016/j.jhydrol.2008.12.017},
 url = {https://www.sciencedirect.com/science/article/pii/S0022169408006355},
 author = {Barbara Tomassetti and Marco Verdecchia and Filippo Giorgi},
 keywords = {Precipitation downscaling, Neural network, Meteorological model}
@@ -156,7 +157,7 @@ title = {The Effect of High-Mountain Asia Topography on Northern Hemisphere Atmo
 journal = {EGU General Assembly 2025, Vienna, Austria, 27 Apr–2 May 2025},
 volume = {EGU25-7283},
 year = {2025},
-doi = {https://doi.org/10.5194/egusphere-egu25-7283},
+doi = {10.5194/egusphere-egu25-7283},
 author = {Mutz, S. G.}
 }
 
