### 2025-09-12 18d4051
tightened paper; reduced wordcount
