### 2025-09-17 #8969 @jatkinson1000
Hi there @ZedThree, @ivan-pi, @anand-me, and @imciner2!

This paper has been submitted to JOSS, the Journal of Open Source Software.
It is a Fortran library (dare I say, 'package') providing statistical models and procedures, and some machine learning functionalities,

Would you be willing to review it?
Please indicate by replying to this thread.

You can find information about reviewing for JOSS [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).
