+++ README.md
@@ -25,7 +25,7 @@ FSML has five thematic modules: Basic statistics (STS), hypothesis tests (TST), 
 
 # Handbook
 
-The [FSML Handbook](http://fsml.mutz.science/page/index.html). includes a short tutorial, detailed API documentation, as well as information for contributors and licence (MIT) details. The documentation pages were generated by [FORD](https://forddocs.readthedocs.io/en/stable/).
+The [FSML Handbook](https://fsml.mutz.science/page/index.html). includes a short tutorial, detailed API documentation, as well as information for contributors and licence (MIT) details. The documentation pages were generated by [FORD](https://forddocs.readthedocs.io/en/stable/).
 
 # Development
 
@@ -44,7 +44,7 @@ The published research below uses some of the to-be-reworked code and demonstrat
 
 ### Progress
 
-Currently covered are procedures for basic statistics (STS), statistical distributions (DST), statistical tests (TST), procedures that rely heavily on linear algebra (LIN), and non-linear algorithmic procedures (NLP). See the full list [here](http://fsml.mutz.science/page/api/index.html). Additionally planned are machine learning framework extensions (e.g., cross-validation) and further additions to the NLP module.
+Currently covered are procedures for basic statistics (STS), statistical distributions (DST), statistical tests (TST), procedures that rely heavily on linear algebra (LIN), and non-linear algorithmic procedures (NLP). See the full list [here](https://fsml.mutz.science/page/api/index.html). Additionally planned are machine learning framework extensions (e.g., cross-validation) and further additions to the NLP module.
 
 # Distribution, Installation, Testing
 
