+++ doc.md
@@ -8,7 +8,7 @@ author: Sebastian G. Mutz
 author_description: Climate/Earth scientist (assoc. professor) with passion for statistics, modelling, AI, games, music, open culture & coding (Fortran & Python). 🇪🇺 🦊🌱
 author_pic: https://avatars.githubusercontent.com/u/46564062?v=4
 github: https://github.com/sebastian-mutz
-website: http://mutz.science/
+website: https://mutz.science/
 email: sebastian@mutz.science
 linkedin: https://www.linkedin.com/in/sebastian-gerhard-m-5b9952210/
 doc_license: by
@@ -57,7 +57,7 @@ FSML has five thematic modules: Basic statistics (STS), hypothesis tests (TST), 
 <br>
 # Handbook
 
-The [FSML Handbook](http://fsml.mutz.science/page/index.html). includes a short tutorial, detailed API documentation, as well as information for contributors and licence (MIT) details. The documentation pages were generated by [FORD](https://forddocs.readthedocs.io/en/stable/).
+The [FSML Handbook](https://fsml.mutz.science/page/index.html). includes a short tutorial, detailed API documentation, as well as information for contributors and licence (MIT) details. The documentation pages were generated by [FORD](https://forddocs.readthedocs.io/en/stable/).
 
 <br>
 # Development
