### 2025-09-27 sebastian-mutz/fsml#12 @ivan-pi
Installation issue and further comments
While following up https://github.com/openjournals/joss-reviews/issues/9058 I tried installing the library using

```txt
$ mkdir install
BADWLRZ-CM11939:fsml di75yas$ fpm install --prefix ./install/
Project is up to date
 + mkdir -p ./install//bin
# Update: build/gfortran_F92076EC7505EC5C/app/fsml -> ./install//bin
```
but the `libfsml.a` and module files were not installed. I guess there is some flag missing from the `fpm.toml`. (I know the library is targeted primarily at `fpm` users, still I'd find it practical if this worked.)

Also, the installation will install the `app/main.f90` as an executable named `fsml`. I'm not sure you really want that.

Regarding the API, is just the `fsml` module meant to be public? I assume it's like that, beacause that file is the one documented. I see some potential issues. For instance `s_dat_read_csv` is public in `fsml_dat`, but is exported from the main module as `fsml_read_csv`. Maybe you could be more explicit about which symbols are part of the public API, so users are aware not to use the names from the internal modules, in case you make a refactoring later on.

I noticed the distribution procedures were implemented using `elemental` core procedures, which are called from "safe" routines that report any invalid arguments. For large arrays, one would typically want to call the core procedure. Is this kind of use-case warranted?

Concerning community guidelines (1. Contribute to the software 2. Report issues or problems with the software 3. Seek support) I noticed you have documents for 1 and 2, but I wasn't sure about 3? Obviously, one cannot support software indefinitely into retirement, or different jobs, so I don't think this can be strictly expected.

Lastly, some typos and other comments:
- L18, strenghts -> strengths
- L27, missing spaces before the citations
- L42, facilitate
- L44, maybe "help counter stagnating ..." could be "promote ..."
- L50-58, the mention of "modules" makes it seem like Fortran modules, maybe you could just say topics or areas?
- L67-70, maybe this is a problem of the template, but the bullet points and code snippets are hard to discern from each other because of the equal indentation; not sure if anything can be done about it
- paragraph 2: in my eyes the statement of need is on lines 24-29 (lack of accessible and documented statistics packages in Fortran), maybe you could slightly rotate the sentences to emphasize this
