### 2025-09-27 2867b81
'fpm install' incl. lib
