### 2025-09-27 248ee73
added support statement
