### 2025-09-27 272a66f
typos and minor text edits
