### 2025-09-27 29dbf3c
clarify only fsml_* are public API
