### 2025-09-27 sebastian-mutz/fsml#12 @sebastian-mutz
Thanks for your input, @ivan-pi.

> but the `libfsml.a` and module files were not installed. I guess there is some flag missing from the `fpm.toml`. (I know the library is targeted primarily at `fpm` users, still I'd find it practical if this worked.)
> Also, the installation will install the `app/main.f90` as an executable named `fsml`. I'm not sure you really want that.

Good point. I updated the `fpm.toml` file, so it now installs `libfsml.a` in `install/lib` ([2867b81](https://github.com/sebastian-mutz/fsml/commit/2867b817227e4f55a5806465028e612f85cffe72)). I use `app/main.f90 `only for quick tests and it does not need to be installed, but as far as I can see, I cannot avoid its installation with the current version of fpm.

> Regarding the API, is just the `fsml` module meant to be public? I assume it's like that, beacause that file is the one documented. I see some potential issues. For instance `s_dat_read_csv` is public in `fsml_dat`, but is exported from the main module as `fsml_read_csv`. Maybe you could be more explicit about which symbols are part of the public API, so users are aware not to use the names from the internal modules, in case you make a refactoring later on.

Correct. Only interfaces in the `fsml` module are part of the public API. For that reason, none of the others are included in the public [API reference](https://fsml.mutz.science/page/api/index.html). I added a note in the `fsml.f90` file (as a FORD "docstring") an an admonition to the [API reference](https://fsml.mutz.science/page/api/index.html) page, noting that only interfaces with the prefix `fsml_` are part of the public API. ([29dbf3c](https://github.com/sebastian-mutz/fsml/commit/29dbf3c96c03ce7e02e0a8c6575ce8b32332bbd1)).

> I noticed the distribution procedures were implemented using `elemental` core procedures, which are called from "safe" routines that report any invalid arguments. For large arrays, one would typically want to call the core procedure. Is this kind of use-case warranted?

Yes, I am currently using the elemental core procedures for large arrays and parallel environments by `use fsml_dst`. However, for the moment, I'm just encouraging the use of "safe" procedures through use of the `fsml` module, because the first users will be (and are already) students still learning to code and working on small projects. Going forward, I will either advise the use of `_core` procedures by importing individual modules (as I'm currently doing) or by bundling all `_core` procedures in something like `fsml_core.f90` as an "unsafe" alternative to `fsml.f90`. That's not decided yet, but the idea is to make sure the `_core` procedures are exposed ~conveniently for those who need them.

> Concerning community guidelines (1. Contribute to the software 2. Report issues or problems with the software 3. Seek support) I noticed you have documents for 1 and 2, but I wasn't sure about 3? Obviously, one cannot support software indefinitely into retirement, or different jobs, so I don't think this can be strictly expected.

I've added a minimalist support statement pointing to the handbook, bug report guidelines, and my e-mail address ([248ee73](https://github.com/sebastian-mutz/fsml/commit/248ee73cacba6f518174f0601b34c27569780079)).

> Lastly, some typos and other comments:
>
>     * L18, strenghts -> strengths
>
>     * L27, missing spaces before the citations
>
>     * L42, facilitate
>
>     * L44, maybe "help counter stagnating ..." could be "promote ..."
>
>     * L50-58, the mention of "modules" makes it seem like Fortran modules, maybe you could just say topics or areas?
>
>     * L67-70, maybe this is a problem of the template, but the bullet points and code snippets are hard to discern from each other because of the equal indentation; not sure if anything can be done about it
>
>     * paragraph 2: in my eyes the statement of need is on lines 24-29 (lack of accessible and documented statistics packages in Fortran), maybe you could slightly rotate the sentences to emphasize this

* I fixed the typos/spaces
* I changed it to "promote" (more concise and positive)
* Simply prepending some spaces in the codeblocks made it easier to discern bullet points from code snippets
* Changed "modules" to "categories" and tightened the sentence somewhat.
* I agree that lines 24-29 describe the need best. Since the sentences that follow go a little into comparisons of similar/related packages (something joss also encourages), I started them as a new paragraph. This way, the second paragraph now ends with lines 24-29, so it doesn't get lost in the middle of a paragraph. ([272a66f](https://github.com/sebastian-mutz/fsml/commit/272a66f6752e83791c4f7b86d395eea900af6aa1))
