### 2025-10-05 sebastian-mutz/fsml#12 @Mikolaj-A-Kowalski
Also related to https://github.com/openjournals/joss-reviews/issues/9058

I am putting it in this issue since the problem I have is also related to the installation instructions. Having no previous experience with `fpm` I have followed [the quick start steps](https://fsml.mutz.science/page/quickstart.html#project-setup) , but on `fpm run` step  it has failed with:
```
fpm build
<ERROR> *cmd_build* Target error: Unable to find source for module dependency: "stdlib_ansi" used by "build/dependencies/fsml/src/fsml_ini.f90"
STOP 1
```
I imagine it is something I am missing in the `fpm` configuration since it does not seem to pull in the `fsml` dependencies, but (once this problem is debugged) it might be a good idea to update the instruction with the step I am missing ;-)
