### 2025-10-05 sebastian-mutz/fsml#12 @sebastian-mutz
Adding the `stdlib` metapackage as a dependency should work (I just tested it). I've updated the quickstart tutorial ([da8d7b2](https://github.com/sebastian-mutz/fsml/commit/da8d7b21fe74a43890588fed1ca460298ab490bd)), re-rendered and uploaded the [handbook section](https://fsml.mutz.science/page/quickstart.html#project-setup).
