### 2025-10-08 sebastian-mutz/fsml#13 @Mikolaj-A-Kowalski
Points from the JOSS review
Related to the JOSS paper review https://github.com/openjournals/joss-reviews/issues/9058

Just wanted to highlight some minor points that it would be good idea to address (although personally I would say they are not required):

- It may be a good idea to include some note of numerical precision in the documentation. What I mean is that `fsml` relies on the intrinsic `sum` of Fortran which (to my knowledge) generally defaults to the naive summation. Hence for some types of data there is a risk of problems related to numerical precision. Including some `note` or a `warning` in the doc to highlight this possibility would make users more careful ;-)
- I have a feeling that the default sentinel value (`-999.0`) often lies in the range of a valid values of the result. Hence it is not possible to detect error with certainty. Perhaps it should be changed to some `NaN` (and `is_sentinel_value` helper function be provided in the API )
- The colouring of the error messages is nice, but sadly it does not work if the output is redirected to a file (The colour characters are included in the output, but they shouldn’t). I fear preserving this functionality in a portable way may be difficult and myself I don't know of the 'correct' way of doing that in Fortran (compilers provide `isatty` function as an extension, but it isn't standard Fortran)
- Perhaps it would be a good  idea to run the test suite under CI (e.g. GitHub Actions)?
- The test suite could be strengthened to handle edge cases and test the error detection (e.g. that functions return sentinel value when input is invalid).
- The link to the Handbook (documentation) webpage could be included in the `About` section in top right. It would help people find the docs more easily
- I am a bit confused about the `dataframe` functionality. It is present in the codebase and the `fsml` module, but does not seem to be tested and documented to the same level as other procedures. Is is part of the public API? Is it an experimental feature in the development?

Also the following points are really just matter of taste so they can be 100% disregarded:
- I thing that the current 'best practices' recommend using more common form of relational operators over the keyword ones (i.e. `>` over `.gt.` etc.)
- Since Fortran 2003 an `allocatable` entity does not need to be explicitly deallocated before it goes out of scope. It will be done by the compiler.
