### 2025-10-08 4726b3d
simple messages (no use of stdlib_ansi colours)
