### 2025-10-08 9826bef
NaN helper functions > replace handling through real sentinels
