+++ src/fsml_dst.f90
@@ -18,6 +18,7 @@ module fsml_dst
   ! load modules
   use :: fsml_ini
   use :: fsml_con
+  use :: fsml_utl
   use :: fsml_err
 
   ! basic options
@@ -82,9 +83,9 @@ impure function f_dst_norm_pdf(x, mu, sigma) result(fx)
 
   ! check if sigma value is valid
   if (sigma_w .le. 0.0_wp) then
-     ! write error message and assign sentinel value if invalid
+     ! write error message and assign NaN value if invalid
      call s_err_print(fsml_error(1))
-     fx = c_sentinel_r
+     fx = f_utl_assign_nan()
      return
   endif
 
@@ -159,18 +160,18 @@ impure function f_dst_norm_cdf(x, mu, sigma, tail) result(p)
 
   ! check if sigma value is valid
   if (sigma_w .le. 0.0_wp) then
-     ! write error message and assign sentinel value if value invalid
+     ! write error message and assign NaN value if value invalid
      call s_err_print(fsml_error(1))
-     p = c_sentinel_r
+     p = f_utl_assign_nan()
      return
   endif
 
   ! check if tail options are valid
   if (tail_w .ne. "left" .and. tail_w .ne. "right" .and. &
      &tail_w .ne. "two" .and. tail_w .ne. "confidence") then
-     ! write error message and assign sentinel value if invalid
+     ! write error message and assign NaN value if invalid
      call s_err_print(fsml_error(2))
-     p = c_sentinel_r
+     p = f_utl_assign_nan()
      return
   endif
 
@@ -265,17 +266,17 @@ impure function f_dst_norm_ppf(p, mu, sigma) result(x)
 
   ! check if sigma value is valid
   if (sigma_w .le. 0.0_wp) then
-     ! write error message and assign sentinel value if invalid
+     ! write error message and assign NaN value if invalid
      call s_err_print(fsml_error(1))
-     x = c_sentinel_r
+     x = f_utl_assign_nan()
      return
   endif
 
   ! check if p value is valid
   if (p .gt. 1.0_wp .or. p .lt. 0.0_wp) then
-     ! write error message and assign sentinel value if invalid
+     ! write error message and assign NaN value if invalid
      call s_err_print(fsml_error(1))
-     x = c_sentinel_r
+     x = f_utl_assign_nan()
      return
   endif
 
@@ -285,7 +286,7 @@ impure function f_dst_norm_ppf(p, mu, sigma) result(x)
   x = f_dst_norm_ppf_core(p, mu_w, sigma_w)
 
   ! issue warning in case of suspicious result
-  if (x .eq. c_sentinel_r) call s_err_warn(fsml_warning(1))
+  if (f_utl_is_nan(x)) call s_err_warn(fsml_warning(1))
 
 end function f_dst_norm_ppf
 
@@ -333,8 +334,8 @@ elemental function f_dst_norm_ppf_core(p, mu, sigma) result(x)
      endif
   enddo
 
-  ! if x not found in iterations, pass sentinel
-  if (i .ge. i_max) x = c_sentinel_r
+  ! if x not found in iterations, pass NaN
+  if (i .ge. i_max) x = f_utl_assign_nan()
 
 end function f_dst_norm_ppf_core
 
@@ -370,17 +371,17 @@ impure function f_dst_t_pdf(x, df, mu, sigma) result(fx)
 
   ! check if sigma value is valid
   if (sigma_w .le. 0.0_wp) then
-     ! write error message and assign sentinel value if invalid
+     ! write error message and assign NaN value if invalid
      call s_err_print(fsml_error(1))
-     fx = c_sentinel_r
+     fx = f_utl_assign_nan()
      return
   endif
 
   ! check if degrees of freedom value is valid
   if (df .le. 0.0_wp) then
[…]
