### 2025-10-08 44350c2
caveat section (mentions naive summation)
