### 2025-10-08 f4335d7
NaN tests for subset of functions
