+++ src/fsml_ini.f90
@@ -19,12 +19,11 @@ module fsml_ini
                               & input_unit, output_unit, error_unit
   ! load module for handling NaNs
   use, intrinsic :: ieee_arithmetic
-  ! load stdlib procedures for ansi terminal colours
-!   use :: stdlib_ansi, only: fg_color_cyan, fg_color_blue, fg_color_magenta, &
-!                           & ansi_code, style_reset, operator(//)
   ! load stblib procedures for linear algebra
   use :: stdlib_linalg, only: eigh, chol
-
+  ! load stdlib procedures for ansi terminal colours
+  ! use :: stdlib_ansi, only: fg_color_cyan, fg_color_blue, fg_color_magenta, &
+  !                         & ansi_code, style_reset, operator(//)
 
 
   ! basic options
@@ -36,8 +35,8 @@ module fsml_ini
   public :: ieee_value, ieee_quiet_nan, ieee_is_nan
   public :: std_i, std_o, std_e, std_rw
   public :: eigh, chol
-!   public :: fg_color_cyan, fg_color_blue, fg_color_magenta, &
-!           & ansi_code, style_reset, operator(//)
+  ! public :: fg_color_cyan, fg_color_blue, fg_color_magenta, &
+  !         & ansi_code, style_reset, operator(//)
 
 ! ==== Declarations
 
