### 2025-10-18 52206be
paper adjustments
