+++ paper/_b/refs.bib
@@ -19,6 +19,48 @@ url = {https://gmd.copernicus.org/articles/16/6479/2023/},
 doi = {10.5194/gmd-16-6479-2023}
 }
 
+@INPROCEEDINGS{jia:2024,
+author={Jia, Yuelong},
+booktitle={2024 3rd International Conference on Artificial Intelligence, Internet of Things and Cloud Computing Technology (AIoTC)},
+title={Analysis of the Impact of Artificial Intelligence on Electricity Consumption},
+year={2024},
+volume={},
+number={},
+pages={57-60},
+keywords={Data centers;Electricity;Computational modeling;Energy conservation;Demand response;Power markets;Energy efficiency;Artificial intelligence;Carbon;Uninterruptible power systems;artificial intelligence;electricity demand;demand response;energy efficiency},
+doi={10.1109/AIoTC63215.2024.10748289}
+}
+
+@inproceedings{dodge:2022,
+author = {Dodge, Jesse and Prewitt, Taylor and Tachet des Combes, Remi and Odmark, Erika and Schwartz, Roy and Strubell, Emma and Luccioni, Alexandra Sasha and Smith, Noah A. and DeCario, Nicole and Buchanan, Will},
+title = {Measuring the Carbon Intensity of AI in Cloud Instances},
+year = {2022},
+isbn = {9781450393522},
+publisher = {Association for Computing Machinery},
+address = {New York, NY, USA},
+url = {https://doi.org/10.1145/3531146.3533234},
+doi = {10.1145/3531146.3533234},
+booktitle = {Proceedings of the 2022 ACM Conference on Fairness, Accountability, and Transparency},
+pages = {1877–1894},
+numpages = {18},
+keywords = {CO2, carbon awareness, carbon intensity, cloud, emissions, grid},
+location = {Seoul, Republic of Korea},
+series = {FAccT '22}
+}
+
+@article{freitag:2021,
+title = {The real climate and transformative impact of ICT: A critique of estimates, trends, and regulations},
+journal = {Patterns},
+volume = {2},
+number = {9},
+pages = {100340},
+year = {2021},
+issn = {2666-3899},
+doi = {10.1016/j.patter.2021.100340},
+url = {https://www.sciencedirect.com/science/article/pii/S2666389921001884},
+author = {Charlotte Freitag and Mike Berners-Lee and Kelly Widdicks and Bran Knowles and Gordon S. Blair and Adrian Friday},
+keywords = {ICT, carbon footprint, trends, AI, big data, data science, IoT, blockchain, policy, regulations},
+}
 
 @article{giorgetta:2018,
 author = {Giorgetta, M. A. and Brokopf, R. and Crueger, T. and Esch, M. and Fiedler, S. and Helmert, J. and Hohenegger, C. and Kornblueh, L. and Köhler, M. and Manzini, E. and Mauritsen, T. and Nam, C. and Raddatz, T. and Rast, S. and Reinert, D. and Sakradzija, M. and Schmidt, H. and Schneck, R. and Schnur, R. and Silvers, L. and Wan, H. and Zängl, G. and Stevens, B.},
@@ -34,6 +76,18 @@ url = {https://agupubs.onlinelibrary.wiley.com/doi/abs/10.1029/2017MS001242},
 eprint = {https://agupubs.onlinelibrary.wiley.com/doi/pdf/10.1029/2017MS001242},
 }
 
+@article{pereira:2021,
+title = {Ranking programming languages by energy efficiency},
+journal = {Science of Computer Programming},
+volume = {205},
+pages = {102609},
+year = {2021},
+issn = {0167-6423},
+doi = {10.1016/j.scico.2021.102609},
[…]
