### 2025-10-18 #9058 @sebastian-mutz
> I feel like it would require much more evidence. That being said, in my opinion it is not really needed since the `FSML` package does introduce extra functionality into Fortran ecosystem and that is sufficient.

Yes, there are very few studies along these lines in general. While less energy efficient languages tend to be used to call more optimised C/C++/Fortran code, widespread casual research practices can add notable overhead. However, that does need much more research and nuanced discussion in separate manuscripts. The point was simply to highlight a potential benefit. However, based on a) everything you mention and b) the main point getting a little lost in the statement of need (pointed out in [issue 12](https://github.com/sebastian-mutz/fsml/issues/12)), I've taken it out.

> I would also recommend to adjust the paper to include the information that at the time of publication `LFortran` is not supported due to its problems with compiling `stdlib`, exactly as it is stated in the online documentation.

I've added that note at the end of the `scope` section (after dependencies are listed). This seems like the most fitting place.

Thank you, @Mikolaj-A-Kowalski . I very much appreciate your effort and interest. :)
