### 2025-10-18 #9058 @ivan-pi
I'll try and finish my review in the next 2-3 days. (I'm traveling 🦘).
