### 2025-11-11 sebastian-mutz/fsml#12 @sebastian-mutz
Just reopened the issue and tagged you, @ivan-pi, as it may help you back into the process. (The open checklist items seem to correspond to points you raised in your original comment here).
