### 2025-12-09 #9058 @jatkinson1000
@sebastian-mutz This is all looking like it is ready for acceptance from the review side.
Thank you @Mikolaj-A-Kowalski and @ivan-pi  for your service!

The next steps are to provide a tagged release for the version of the code that was reviewed, and place this same version in an online archive with a DOI (e.g. Zenodo, figshare or an institution archive etc.).
Please ensure that the archival deposit has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCiD.

Once you have done this please could you post the version number and archive DOI here please.

In the meantime I will take a proof-read of the paper to try and catch any small final issues from the editorial side.
