### 2025-12-09 #9058 @jatkinson1000
@sebastian-mutz When I generate the pdf the result does not seem to match up with the latest version in the repository.

Please could you check that you can build locally or using the [JOSS GitHub Action](https://github.com/marketplace/actions/open-journals-pdf-generator) as a first check.
In the meantime I will also look to see if it is an issue with the bot our end.
