### 2025-12-09 a00a8d2
removed paper.md backup
