+++ paper/_b/refs.bib
@@ -1,218 +1,0 @@
-@article{lang:2024,
-  title={AIFS-ECMWF's data-driven forecasting system},
-  author={Lang, Simon and Alexe, Mihai and Chantry, Matthew and Dramsch, Jesper and Pinault, Florian and Raoult, Baudouin and Clare, Mariana CA and Lessig, Christian and Maier-Gerber, Michael and Magnusson, Linus and others},
-  journal={arXiv preprint arXiv:2406.01465},
-  year={2024},
-  doi = {10.48550/arXiv.2406.01465}
-}
-
-@article{boateng:2023,
-author = {Boateng, D. and Mutz, S. G.},
-title = {pyESDv1.0.1: an open-source Python framework for empirical-statistical
-downscaling of climate information},
-journal = {Geoscientific Model Development},
-volume = {16},
-year = {2023},
-number = {22},
-pages = {6479--6514},
-url = {https://gmd.copernicus.org/articles/16/6479/2023/},
-doi = {10.5194/gmd-16-6479-2023}
-}
-
-@INPROCEEDINGS{jia:2024,
-author={Jia, Yuelong},
-booktitle={2024 3rd International Conference on Artificial Intelligence, Internet of Things and Cloud Computing Technology (AIoTC)},
-title={Analysis of the Impact of Artificial Intelligence on Electricity Consumption},
-year={2024},
-volume={},
-number={},
-pages={57-60},
-keywords={Data centers;Electricity;Computational modeling;Energy conservation;Demand response;Power markets;Energy efficiency;Artificial intelligence;Carbon;Uninterruptible power systems;artificial intelligence;electricity demand;demand response;energy efficiency},
-doi={10.1109/AIoTC63215.2024.10748289}
-}
-
-@inproceedings{dodge:2022,
-author = {Dodge, Jesse and Prewitt, Taylor and Tachet des Combes, Remi and Odmark, Erika and Schwartz, Roy and Strubell, Emma and Luccioni, Alexandra Sasha and Smith, Noah A. and DeCario, Nicole and Buchanan, Will},
-title = {Measuring the Carbon Intensity of AI in Cloud Instances},
-year = {2022},
-isbn = {9781450393522},
-publisher = {Association for Computing Machinery},
-address = {New York, NY, USA},
-url = {https://doi.org/10.1145/3531146.3533234},
-doi = {10.1145/3531146.3533234},
-booktitle = {Proceedings of the 2022 ACM Conference on Fairness, Accountability, and Transparency},
-pages = {1877–1894},
-numpages = {18},
-keywords = {CO2, carbon awareness, carbon intensity, cloud, emissions, grid},
-location = {Seoul, Republic of Korea},
-series = {FAccT '22}
-}
-
-@article{freitag:2021,
-title = {The real climate and transformative impact of ICT: A critique of estimates, trends, and regulations},
-journal = {Patterns},
-volume = {2},
-number = {9},
-pages = {100340},
-year = {2021},
-issn = {2666-3899},
-doi = {10.1016/j.patter.2021.100340},
-url = {https://www.sciencedirect.com/science/article/pii/S2666389921001884},
-author = {Charlotte Freitag and Mike Berners-Lee and Kelly Widdicks and Bran Knowles and Gordon S. Blair and Adrian Friday},
-keywords = {ICT, carbon footprint, trends, AI, big data, data science, IoT, blockchain, policy, regulations},
-}
-
-@article{giorgetta:2018,
[…]
