### 2025-12-20 9942442
JOSS review; fixed capitalisation in bibliography
