+++ paper/refs.bib
@@ -8,7 +8,7 @@
 
 @article{boateng:2023,
 author = {Boateng, D. and Mutz, S. G.},
-title = {pyESDv1.0.1: an open-source Python framework for empirical-statistical
+title = {pyESDv1.0.1: an open-source {P}ython framework for empirical-statistical
 downscaling of climate information},
 journal = {Geoscientific Model Development},
 volume = {16},
@@ -22,7 +22,7 @@ doi = {10.5194/gmd-16-6479-2023}
 
 @article{giorgetta:2018,
 author = {Giorgetta, M. A. and Brokopf, R. and Crueger, T. and Esch, M. and Fiedler, S. and Helmert, J. and Hohenegger, C. and Kornblueh, L. and Köhler, M. and Manzini, E. and Mauritsen, T. and Nam, C. and Raddatz, T. and Rast, S. and Reinert, D. and Sakradzija, M. and Schmidt, H. and Schneck, R. and Schnur, R. and Silvers, L. and Wan, H. and Zängl, G. and Stevens, B.},
-title = {ICON-A, the Atmosphere Component of the ICON Earth System Model: I. Model Description},
+title = {ICON-A, the Atmosphere Component of the ICON {E}arth System Model: I. Model Description},
 journal = {Journal of Advances in Modeling Earth Systems},
 volume = {10},
 number = {7},
@@ -38,7 +38,7 @@ eprint = {https://agupubs.onlinelibrary.wiley.com/doi/pdf/10.1029/2017MS001242},
 @ARTICLE{kedward:2022,
   author={Kedward, Laurence J. and Aradi, Bálint and Čertík, Ondřej and Curcic, Milan and Ehlert, Sebastian and Engel, Philipp and Goswami, Rohit and Hirsch, Michael and Lozada-Blanco, Asdrubal and Magnin, Vincent and Markus, Arjen and Pagone, Emanuele and Pribec, Ivan and Richardson, Brad and Snyder, Harris and Urban, John and Vandenplas, Jérémie},
   journal={Computing in Science & Engineering},
-  title={The State of Fortran},
+  title={The State of {F}ortran},
   year={2022},
   volume={24},
   number={2},
@@ -99,7 +99,7 @@ keywords = {Precipitation downscaling, Neural network, Meteorological model}
 }
 
 @article{mutz:2025,
-title = {The Effect of High-Mountain Asia Topography on Northern Hemisphere Atmospheric Flow},
+title = {The Effect of High-Mountain {A}sia Topography on {N}orthern {H}emisphere Atmospheric Flow},
 journal = {EGU General Assembly 2025, Vienna, Austria, 27 Apr–2 May 2025},
 volume = {EGU25-7283},
 year = {2025},
@@ -109,7 +109,7 @@ author = {Mutz, S. G.}
 
 @article{mutz:2018,
 AUTHOR = {Mutz, S. G. and Ehlers, T. A. and Werner, M. and Lohmann, G. and Stepanek, C. and Li, J.},
-TITLE = {Estimates of late Cenozoic climate change relevant to Earth surface
+TITLE = {Estimates of {L}ate {C}enozoic climate change relevant to {E}arth surface
 processes in tectonically active orogens},
 JOURNAL = {Earth Surface Dynamics},
 VOLUME = {6},
@@ -122,8 +122,8 @@ DOI = {10.5194/esurf-6-271-2018}
 
 @article{mutz:2019,
 AUTHOR = {Mutz, S. G. and Ehlers, T. A.},
-TITLE = {Detection and explanation of spatiotemporal patterns in Late Cenozoic
-palaeoclimate change relevant to Earth surface processes},
+TITLE = {Detection and explanation of spatiotemporal patterns in {L}ate {C}enozoic
[…]
