+++ CITATION.cff
@@ -1,0 +1,28 @@
+cff-version: "1.2.0"
+authors:
+- family-names: Mutz
+  given-names: Sebastian Gerhard
+  orcid: "https://orcid.org/0000-0001-8180-6150"
+doi: 10.5281/zenodo.17872189
+message: If you use this software, please cite our article in the
+  Journal of Open Source Software.
+preferred-citation:
+  authors:
+  - family-names: Mutz
+    given-names: Sebastian Gerhard
+    orcid: "https://orcid.org/0000-0001-8180-6150"
+  date-published: 2025-12-20
+  doi: 10.21105/joss.09058
+  issn: 2475-9066
+  issue: 116
+  journal: Journal of Open Source Software
+  publisher:
+    name: Open Journals
+  start: 9058
+  title: FSML -- A Modern Fortran Statistics and Machine Learning
+    Library
+  type: article
+  url: "https://joss.theoj.org/papers/10.21105/joss.09058"
+  volume: 10
+title: FSML -- A Modern Fortran Statistics and Machine Learning Library
+
