### 2025-12-20 149d65a
bibtex reformat
