+++ README.md
@@ -58,15 +58,16 @@ Mutz, S. G., (2025). FSML – A Modern Fortran Statistics and Machine Learning L
 **BibTex**:
 ```bibtex
 @article{Mutz2025,
-doi = {10.21105/joss.09058},
-url = {https://doi.org/10.21105/joss.09058},
+author = {Mutz, Sebastian Gerhard},
+title = {FSML – A Modern Fortran Statistics and Machine Learning Library},
 year = {2025},
+journal = {Journal of Open Source Software},
 publisher = {The Open Journal},
 volume = {10},
 number = {116},
 pages = {9058},
-author = {Mutz, Sebastian Gerhard},
-title = {FSML – A Modern Fortran Statistics and Machine Learning Library}, journal = {Journal of Open Source Software} }
+doi = {10.21105/joss.09058},
+url = {https://doi.org/10.21105/joss.09058} }
 ```
 
 
