+++ doc/citation.md
@@ -1,0 +1,25 @@
+---
+title: Citation
+---
+
+# Reference
+
+If you use `FSML` for your project, please use the following reference for citation:
+
+Mutz, S. G., (2025). FSML – A Modern Fortran Statistics and Machine Learning Library. Journal of Open Source Software, 10(116), 9058, [https://doi.org/10.21105/joss.09058](https://doi.org/10.21105/joss.09058)
+
+# BibTex
+
+```bibtex
+@article{Mutz2025,
+author = {Mutz, Sebastian Gerhard},
+title = {FSML – A Modern Fortran Statistics and Machine Learning Library},
+year = {2025},
+journal = {Journal of Open Source Software},
+publisher = {The Open Journal},
+volume = {10},
+number = {116},
+pages = {9058},
+doi = {10.21105/joss.09058},
+url = {https://doi.org/10.21105/joss.09058} }
+```
