+++ doc/index.md
@@ -14,6 +14,10 @@ A complete list of public-facing procedures can be found [here](./api/index.html
 
 If you're interested in contributing to the project, please read the contributor guidelines [here](./contributing.html).
 
+# Citation
+
+If you use this library for your project, please cite it as shown [here](./citation.html).
+
 # Licence
 
 View the licence (MIT) information [here](./licence.html).
