+++ doc.md
@@ -1,6 +1,6 @@
 ---
 project: FSML
-summary: FSML (Fortran Statistics and Machine Learning) is a scientific toolkit for statistics and machine learning - from basic correlation and statistical hypothesis testing to different multivariate regression and clustering procedures. It is written in modern Fortran (2008+) and offered as an [FPM](https://fpm.fortran-lang.org/) package for easy integration into your projects. The source code is hosted on [GitHub](https://github.com/sebastian-mutz/fsml) and released under the [MIT licence](./page/licence.html).
+summary: FSML (Fortran Statistics and Machine Learning) is a scientific toolkit for statistics and machine learning - from basic correlation and statistical hypothesis testing to different multivariate regression and clustering procedures. It is written in modern Fortran (2008+) and offered as an [FPM](https://fpm.fortran-lang.org/) package for easy integration into your projects. The source code is hosted on [GitHub](https://github.com/sebastian-mutz/fsml) and released under the [MIT licence](./page/licence.html). The associated paper can be accessed [here](https://doi.org/10.21105/joss.09058).
 license: MIT license
 project_download: https://github.com/sebastian-mutz/fsml/archive/refs/heads/master.zip
 favicon: ./doc/media/logo/FSML_icon.png
