### 2025-12-21 9940671
added release archive + notes to repo
