+++ releases/v0.1.0/notes.md
@@ -1,0 +1,15 @@
+# FSML v0.1.0 (alpha) – initial release
+ 
+This is the initial release of the FSML (Fortran Statistics and Machine Learning) library/package. This version was reviewed in the process of publishing its associated paper in the _Journal of Open Source Software (JOSS)_.
+
+## Publication in JOSS
+
+The associated paper (_FSML – A Modern Fortran Statistics and Machine Learning Library_) was peer-reviewed and published in JOSS. It can be accessed [here](https://doi.org/10.21105/joss.09058). As part of the publication process, this version (v0.1.0) was also archives on Zenodo [here](https://doi.org/10.5281/zenodo.17872189).
+
+## Documentation and Tutorial
+
+A tutorial, examples, and detailed API documentation is hosted on [fsml.mutz.science](https://fsml.mutz.science/).
+
+## Blog Post
+
+This [blog post](https://sebastianmutz.eu/blog/posts/2025-12-20/) provides a little more context, notes on some design choices, and where `FSML` fits into the Fortran statistics and machine learning ecosystem.
