### 2025-12-22 d84d459
Compilable Lasso implementation (no validation yet).
