+++ src/fsml_lin.f90
@@ -766,6 +766,203 @@ pure function f_lin_mahalanobis_core(x, y, cov) result(dist)
 
 end function f_lin_mahalanobis_core
 
+subroutine s_lin_lasso(x, y, nd, nv, lambda, b0, b, r2, y_hat, se, cov_b, rho, tol, maxiter)
 
+!  ==== Description
+!!  Least Absolute Shrinkage and Selection Operator with intercept.
+!!
+!! Computes:
+!!
+!! Notes:
+
+! ==== Declarations
+  integer(i4), intent(in)            :: nd
+  integer(i4), intent(in)            :: nv
+  real(wp)   , intent(in)            :: x(nd, nv)
+  real(wp)   , intent(in)            :: y(nd)
+  real(wp)   , intent(in)            :: lambda
+  real(wp)   , intent(out)           :: b0
+  real(wp)   , intent(out)           :: b(nv)
+  real(wp)   , intent(out)           :: r2
+  real(wp)   , intent(out), optional :: y_hat(nd)
+  real(wp)   , intent(out), optional :: se(nv)
+  real(wp)   , intent(out), optional :: cov_b(nv, nv)
+  real(wp)   , intent(in) , optional :: rho
+  real(wp)   , intent(in) , optional :: tol
+  integer(i4), intent(in) , optional :: maxiter
+  ! ---- Internal variables
+  real(wp)                           :: rho_
+  real(wp)                           :: xc(nv, nv), yc(nd)
+  real(wp)                           :: xt(nv, nd)
+  real(wp)                           :: xtx(nv, nv)
+  real(wp)                           :: xty(nv)
+  real(wp)                           :: P(nv, nv)
+  real(wp)                           :: res(nd)
+  real(wp)                           :: sse
+  real(wp)                           :: sst
+  real(wp)                           :: y_bar
+  integer(i4)                        :: i, j
+! ==== Instructions
+
+  ! ---- validate input
+  if (nd <= nv + 1) then
+    call s_err_print("[fsml error] Lasso: Number of obersvations must&
+                    & exceed number of predictors + intercept.")
+    error stop
+  endif
+  if (lambda < 0.0_wp) then
+      call s_err_print("[fsml error] Lasso: lambda must be non-zero positive.")
+      error stop
+  endif
+
+  ! ---- optional arguments.
+  rho_ = 50.0_wp; if (present(rho)) rho_ = rho
+
+  ! ---- Preprocessing
+
+  ! center dependent variable.
+  yc = y - f_sts_mean_core(y)
+
+  ! center regressors.
+  do concurrent(j=1:nv)
+      xc(:, j) = x(:, j) - f_sts_mean_core(x(:, j))
+  enddo
+
+  ! ---- Cholesky factorization of P = rho*I + XᵗX
+
+  ! compute transposed matrix and XᵗX
+  xt  = transpose(xc)
+  xtx = matmul(xt, xc)
+
+  ! compute P = rho*I + XᵗX.
+  do concurrent(i=1:nv, j=1:nv)
+      P(i, j) = merge(rho_ + xtx(i, j), xtx(i, j), i == j)
+  enddo
+
+  ! ---- compute coefficients: full vector including intercept
+  xty = matmul(xt, yc)
+  b0  = 0.0_wp
+  b   = 0.0_wp
+
+  ! get intercept coefficient.
+  b0 = f_sts_mean(y)
+
+  ! get predictor coefficients.
+  b = admm(nv, P, xty, lambda, rho_, maxiter, tol)
+
+  ! ---- predicted values and residuals.
+  if (present(y_hat)) then
[…]
