### 2025-12-22 ac91503
In-code documentation.
