+++ src/fsml_lin.f90
@@ -772,36 +772,45 @@ subroutine s_lin_lasso(x, y, nd, nv, lambda, b0, b, r2, y_hat, se, cov_b, rho, t
 !!  Least Absolute Shrinkage and Selection Operator with intercept.
 !!
 !! Computes:
+!!  - Intercept b0 (scalar)
+!!  - Predictor coefficients b(nv)
+!!  - Coefficient of determination R²
+!!  - Standard errors se(nv) (lasso-adjusted)
+!!  - Covariance matrix of predictors cov_b(nv, nv) (lasso-adjusted)
 !!
 !! Notes:
+!!  - When lambda (λ) = 0, this reduces to ordinary least squares (OLS).
+!!  - Lasso problem is solved using the Alternating Direction Method of Multipliers.
 
 ! ==== Declarations
-  integer(i4), intent(in)            :: nd
-  integer(i4), intent(in)            :: nv
-  real(wp)   , intent(in)            :: x(nd, nv)
-  real(wp)   , intent(in)            :: y(nd)
-  real(wp)   , intent(in)            :: lambda
-  real(wp)   , intent(out)           :: b0
-  real(wp)   , intent(out)           :: b(nv)
-  real(wp)   , intent(out)           :: r2
-  real(wp)   , intent(out), optional :: y_hat(nd)
-  real(wp)   , intent(out), optional :: se(nv)
-  real(wp)   , intent(out), optional :: cov_b(nv, nv)
-  real(wp)   , intent(in) , optional :: rho
-  real(wp)   , intent(in) , optional :: tol
-  integer(i4), intent(in) , optional :: maxiter
+  integer(i4), intent(in)            :: nd            !! number of datapoints
+  integer(i4), intent(in)            :: nv            !! number of predictors (excluding intercept)
+  real(wp)   , intent(in)            :: x(nd, nv)     !! predictor data matrix (no intercept column)
+  real(wp)   , intent(in)            :: y(nd)         !! response vector
+  real(wp)   , intent(in)            :: lambda        !! lasso penaly parameter (≥ 0, non-optional)
+
+  real(wp)   , intent(out)           :: b0            !! intercept coefficient
+  real(wp)   , intent(out)           :: b(nv)         !! predictor coefficients
+  real(wp)   , intent(out)           :: r2            !! coefficient of determination R²
+  real(wp)   , intent(out), optional :: y_hat(nd)     !! predicted y values
+  real(wp)   , intent(out), optional :: se(nv)        !! standard errors of predictor coefficients
+  real(wp)   , intent(out), optional :: cov_b(nv, nv) !! covariance matrix of predictor coefficients
+  real(wp)   , intent(in) , optional :: rho           !! ADMM weight for augmented Lagrangian
+  real(wp)   , intent(in) , optional :: tol           !! Tolerance for ADMM solver
+  integer(i4), intent(in) , optional :: maxiter       !! Maximum number of iterations for ADMM
   ! ---- Internal variables
-  real(wp)                           :: rho_
-  real(wp)                           :: xc(nv, nv), yc(nd)
-  real(wp)                           :: xt(nv, nd)
-  real(wp)                           :: xtx(nv, nv)
-  real(wp)                           :: xty(nv)
-  real(wp)                           :: P(nv, nv)
-  real(wp)                           :: res(nd)
-  real(wp)                           :: sse
-  real(wp)                           :: sst
-  real(wp)                           :: y_bar
-  integer(i4)                        :: i, j
+  real(wp)                           :: rho_          !! ADMM weights for augmented Lagrangian
+  real(wp)                           :: xc(nv, nv)    !! centered predictor data matrix
+  real(wp)                           :: yc(nd)        !! centered response vector
[…]
