### 2025-12-22 c4a1832
Purified function admm.
