### 2025-12-22 35f6938
Code runs (not validated yet).
