### 2025-12-22 45d8a64
Fixed implementation.
