+++ src/fsml_lin.f90
@@ -800,12 +800,11 @@ subroutine s_lin_lasso(x, y, nd, nv, lambda, b0, b, r2, y_hat, se, cov_b, rho, t
   integer(i4), intent(in) , optional :: maxiter       !! Maximum number of iterations for ADMM
   ! ---- Internal variables
   real(wp)                           :: rho_          !! ADMM weights for augmented Lagrangian
-  real(wp)                           :: xc(nd, nv)    !! centered predictor data matrix
-  real(wp)                           :: yc(nd)        !! centered response vector
-  real(wp)                           :: xt(nv, nd)    !! transpose of xc
-  real(wp)                           :: xtx(nv, nv)   !! XᵀX matrix
-  real(wp)                           :: xty(nv)       !! Xᵀy
-  real(wp)                           :: P(nv, nv)     !! Symmetric positive definite matrix.
+  real(wp)                           :: x1(nd, nv+1)    !! centered predictor data matrix
+  real(wp)                           :: xt(nv+1, nd)    !! transpose of xc
+  real(wp)                           :: xtx(nv+1, nv+1)   !! XᵀX matrix
+  real(wp)                           :: xty(nv+1)       !! Xᵀy
+  real(wp)                           :: coefs(nv+1)
   real(wp)                           :: res(nd)       !! residuals
   real(wp)                           :: sse           !! sum of squared errors
   real(wp)                           :: sst           !! total sum of squares
@@ -829,38 +828,38 @@ subroutine s_lin_lasso(x, y, nd, nv, lambda, b0, b, r2, y_hat, se, cov_b, rho, t
 
   ! ---- Preprocessing
 
-  ! center dependent variable.
-  yc = y - f_sts_mean_core(y)
-
   ! center regressors.
-  do concurrent(j=1:nv)
-      xc(:, j) = x(:, j) - f_sts_mean_core(x(:, j))
+  ! x1(:, 1) = 1.0_wp; x1(:, 2:) = x
+  do i = 1, nd
+     x1(i,1) = 1.0_wp
+     do j = 1, nv
+        x1(i,j+1) = x(i,j)
+     enddo
   enddo
 
   ! ---- Cholesky factorization of P = rho*I + XᵗX
 
   ! compute transposed matrix and XᵗX
-  xt  = transpose(xc)
-  xtx = matmul(xt, xc)
-
+  xt  = transpose(x1)
+  xtx = matmul(xt, x1)
 
   ! ---- compute coefficients: full vector including intercept
-  xty = matmul(xt, yc)
+  xty = matmul(xt, y)
   b0  = 0.0_wp
   b   = 0.0_wp
 
-  ! get intercept coefficient.
-  b0 = f_sts_mean(y)
+  ! solve lasso problem.
+  coefs = admm(nv+1, xtx, xty, lambda, rho_, maxiter, tol)
 
-  ! get predictor coefficients.
-  b = admm(nv, xtx, xty, lambda, rho_, maxiter, tol)
+  ! get coefficients.
+  b0 = coefs(1) ; b = coefs(2:)
 
   ! ---- predicted values and residuals.
   if (present(y_hat)) then
-    y_hat = b0 + matmul(xc, b)
+    y_hat = b0 + matmul(x, b)
     res = y - y_hat
   else
-    res = y - (b0 + matmul(xc, b))
+    res = y - (b0 + matmul(x, b))
   endif
 
   ! ---- R² calculation.
@@ -879,7 +878,7 @@ subroutine s_lin_lasso(x, y, nd, nv, lambda, b0, b, r2, y_hat, se, cov_b, rho, t
 
 end subroutine s_lin_lasso
 
-pure function admm(n, xtx, xty, lambda, rho, maxiter, tol) result(x)
[…]
