+++ src/fsml_lin.f90
@@ -783,33 +783,30 @@ subroutine s_lin_lasso(x, y, nd, nv, lambda, b0, b, r2, y_hat, se, cov_b, rho, t
 !!  - Lasso problem is solved using the Alternating Direction Method of Multipliers.
 
 ! ==== Declarations
-  integer(i4), intent(in)            :: nd            !! number of datapoints
-  integer(i4), intent(in)            :: nv            !! number of predictors (excluding intercept)
-  real(wp)   , intent(in)            :: x(nd, nv)     !! predictor data matrix (no intercept column)
-  real(wp)   , intent(in)            :: y(nd)         !! response vector
-  real(wp)   , intent(in)            :: lambda        !! lasso penaly parameter (≥ 0, non-optional)
-
-  real(wp)   , intent(out)           :: b0            !! intercept coefficient
-  real(wp)   , intent(out)           :: b(nv)         !! predictor coefficients
-  real(wp)   , intent(out)           :: r2            !! coefficient of determination R²
-  real(wp)   , intent(out), optional :: y_hat(nd)     !! predicted y values
-  real(wp)   , intent(out), optional :: se(nv)        !! standard errors of predictor coefficients
-  real(wp)   , intent(out), optional :: cov_b(nv, nv) !! covariance matrix of predictor coefficients
-  real(wp)   , intent(in) , optional :: rho           !! ADMM weight for augmented Lagrangian
-  real(wp)   , intent(in) , optional :: tol           !! Tolerance for ADMM solver
-  integer(i4), intent(in) , optional :: maxiter       !! Maximum number of iterations for ADMM
+  integer(i4), intent(in)            :: nd              !! number of datapoints
+  integer(i4), intent(in)            :: nv              !! number of predictors (excluding intercept)
+  real(wp)   , intent(in)            :: x(nd, nv)       !! predictor data matrix (no intercept column)
+  real(wp)   , intent(in)            :: y(nd)           !! response vector
+  real(wp)   , intent(in)            :: lambda          !! lasso penaly parameter (≥ 0, non-optional)
+  real(wp)   , intent(out)           :: b0              !! intercept coefficient
+  real(wp)   , intent(out)           :: b(nv)           !! predictor coefficients
+  real(wp)   , intent(out)           :: r2              !! coefficient of determination R²
+  real(wp)   , intent(out), optional :: y_hat(nd)       !! predicted y values
+  real(wp)   , intent(out), optional :: se(nv)          !! standard errors of predictor coefficients
+  real(wp)   , intent(out), optional :: cov_b(nv, nv)   !! covariance matrix of predictor coefficients
+  real(wp)   , intent(in) , optional :: rho             !! ADMM weight for augmented Lagrangian
+  real(wp)   , intent(in) , optional :: tol             !! Tolerance for ADMM solver
+  integer(i4), intent(in) , optional :: maxiter         !! Maximum number of iterations for ADMM
   ! ---- Internal variables
-  real(wp)                           :: rho_          !! ADMM weights for augmented Lagrangian
   real(wp)                           :: x1(nd, nv+1)    !! centered predictor data matrix
   real(wp)                           :: xt(nv+1, nd)    !! transpose of xc
-  real(wp)                           :: xtx(nv+1, nv+1)   !! XᵀX matrix
+  real(wp)                           :: xtx(nv+1, nv+1) !! XᵀX matrix
[…]
