### 2025-12-22 442f12c
Change default admm weight from 100 to 50.
