### 2025-12-23 8dde29f
In-place Cholesky + over-relaxation ADMM
