### 2025-12-23 e2324d8
Almost complete specifications (only covariance expression missing).
