### 2026-01-17 468dd36
minor doc/readme update
