+++ releases/draft_v0.1.1/notes.md
@@ -1,0 +1,28 @@
+# FSML v0.1.1
+ 
+## Changelog
+The following has changed since the last release (v0.1.0):
+
+### LIN module
+- [ ] Added Least Absolute Shrinkage and Selection Operator (LASSO) regression [contributed by @loiseaujc].
+- [ ] Added Canonical Correlation Analysis (CCA) [contributed by @loiseaujc].
+
+### DST module
+- [ ] Added log-logistic distribution PDF, CDF, and PPF.
+- [ ] Added logistic distribution PDF, CDF, and PPF.
+
+### NLP module
+
+- [ ] Implemented Lance-Williams algorithm for clustering; more efficient than re-computing Mahalanobis distance
+
+## New Contributors
+- @loiseaujc/[Jean-Christophe Loiseau](https://loiseaujc.github.io/) (PRs).
+
+## Communication
+
+- [ ] Conference Presentation: FSML's development approach was presented at the EGU 2026 General Assembly session ["Scientific software development in the Geosciences: good practices, pitfalls and solutions"](https://www.egu26.eu/session/57206) to encourage more sustainable development practices in the scientific community.
+- [ ] Blog post: This blog post is a brief development update, providing more context and background.
+
+## Application Spotlight
+- @Khadar146/Khadar Dahir Abdisalan used FSML's distribution functions to statistically model future droughts in Somaliland.
+
