### 2026-01-17 3f81edb
logistic PDF, CDF, PPF
