+++ src/fsml_dst.f90
@@ -47,6 +47,9 @@ module fsml_dst
   public :: f_dst_gpd_pdf, f_dst_gpd_pdf_core
   public :: f_dst_gpd_cdf, f_dst_gpd_cdf_core
   public :: f_dst_gpd_ppf, f_dst_gpd_ppf_core
+  public :: f_dst_logistic_pdf, f_dst_logistic_pdf_core
+  public :: f_dst_logistic_cdf, f_dst_logistic_cdf_core
+  public :: f_dst_logistic_ppf, f_dst_logistic_ppf_core
 
   ! additional core features; only exposed in fsml_dst
   public :: f_dst_gammai_core, f_dst_betai_core
@@ -245,12 +248,12 @@ impure function f_dst_norm_ppf(p, mu, sigma) result(x)
 !! Handles optional arguments and invalid values for arguments.
 
 ! ==== Declarations
-  real(wp)   , intent(in)           :: p                !! probability between 0.0 - 1.0
-  real(wp)   , intent(in), optional :: mu               !! distribution location (mean)
-  real(wp)   , intent(in), optional :: sigma            !! distribution dispersion/scale (standard deviation)
-  real(wp)                          :: mu_w             !! final value of mu
-  real(wp)                          :: sigma_w          !! final value of sigma
-  real(wp)                          :: x                !! sample position
+  real(wp)   , intent(in)           :: p       !! probability between 0.0 - 1.0
+  real(wp)   , intent(in), optional :: mu      !! distribution location (mean)
+  real(wp)   , intent(in), optional :: sigma   !! distribution dispersion/scale (standard deviation)
+  real(wp)                          :: mu_w    !! final value of mu
+  real(wp)                          :: sigma_w !! final value of sigma
+  real(wp)                          :: x       !! sample position
 
 ! ==== Instructions
 
@@ -2235,6 +2238,274 @@ elemental function f_dst_gpd_ppf_core(p, xi, mu, sigma) result(x)
   endif
 
 end function f_dst_gpd_ppf_core
+
+
+! ==================================================================== !
+! -------------------------------------------------------------------- !
+impure function f_dst_logistic_pdf(x, mu, scale) result(fx)
+
+! ==== Description
+!! Impure wrapper function for `f_dst_logis_pdf_core`.
+!! Handles optional arguments and invalid values for arguments.
+
+! ==== Declarations
+  real(wp), intent(in)           :: x       !! sample position
+  real(wp), intent(in), optional :: mu      !! distribution location
+  real(wp), intent(in), optional :: scale   !! distribution scale
+  real(wp)                       :: mu_w    !! final value for mu
+  real(wp)                       :: scale_w !! final value for scale
+  real(wp)                       :: fx
+
+! ==== Instructions
+
+! ---- handle input
+
+  ! assume location/mean = 0, overwrite if specified
+  mu_w = 0.0_wp
+  if (present(mu)) mu_w = mu
+
+  ! assume sscale = 1, overwrite if specified
+  scale_w = 1.0_wp
+  if (present(scale)) scale_w = scale
+
+  ! check if scale value is valid
+  if (scale_w .le. 0.0_wp) then
+     call s_err_print(fsml_error(1))
+     fx = f_utl_assign_nan()
+     return
+  endif
+
+! ---- compute PDF
+
+  ! call pure function to calculate probability/fx
+  fx = f_dst_logistic_pdf_core(x, mu_w, scale_w)
+
+end function f_dst_logistic_pdf
+
+
+! ==================================================================== !
+! -------------------------------------------------------------------- !
+elemental function f_dst_logistic_pdf_core(x, mu, scale) result(fx)
+
+! ==== Description
[…]
