### 2026-01-18 3ad09b9
log logistic PDF, CDF, PPF
