+++ src/fsml_dst.f90
@@ -50,6 +50,9 @@ module fsml_dst
   public :: f_dst_logistic_pdf, f_dst_logistic_pdf_core
   public :: f_dst_logistic_cdf, f_dst_logistic_cdf_core
   public :: f_dst_logistic_ppf, f_dst_logistic_ppf_core
+  public :: f_dst_llogistic_pdf, f_dst_llogistic_pdf_core
+  public :: f_dst_llogistic_cdf, f_dst_llogistic_cdf_core
+  public :: f_dst_llogistic_ppf, f_dst_llogistic_ppf_core
 
   ! additional core features; only exposed in fsml_dst
   public :: f_dst_gammai_core, f_dst_betai_core
@@ -2245,7 +2248,7 @@ end function f_dst_gpd_ppf_core
 impure function f_dst_logistic_pdf(x, mu, scale) result(fx)
 
 ! ==== Description
-!! Impure wrapper function for `f_dst_logis_pdf_core`.
+!! Impure wrapper function for `f_dst_logistic_pdf_core`.
 !! Handles optional arguments and invalid values for arguments.
 
 ! ==== Declarations
@@ -2305,7 +2308,7 @@ elemental function f_dst_logistic_pdf_core(x, mu, scale) result(fx)
 
   ! CDF evaluation
   if (z .ge. 0.0_wp) then
-    p = 1.0_wp / (1.0_wp + exp(-z))
+     p = 1.0_wp / (1.0_wp + exp(-z))
   else
      p = exp(z) / (1.0_wp + exp(z))
   endif
@@ -2325,7 +2328,7 @@ end function f_dst_logistic_pdf_core
 impure function f_dst_logistic_cdf(x, mu, scale, tail) result(p)
 
 ! ==== Description
-!! Impure wrapper function for `f_dst_logis_cdf_core`.
+!! Impure wrapper function for `f_dst_logistic_cdf_core`.
 !! Handles optional arguments and invalid values for arguments.
 
 ! ==== Declarations
@@ -2435,7 +2438,7 @@ end function f_dst_logistic_cdf_core
 impure function f_dst_logistic_ppf(p, mu, scale) result(x)
 
 ! ==== Description
-!! Impure wrapper function for `f_dst_logis_ppf_core`.
+!! Impure wrapper function for `f_dst_logistic_ppf_core`.
 !! Handles optional arguments and invalid values for arguments.
 
 ! ==== Declarations
@@ -2506,6 +2509,294 @@ elemental function f_dst_logistic_ppf_core(p, mu, scale) result(x)
   x = mu + scale * log(p / (1.0_wp - p))
 
 end function f_dst_logistic_ppf_core
+
+
+! ==================================================================== !
+! -------------------------------------------------------------------- !
+impure function f_dst_llogistic_pdf(x, mu, alpha, beta) result(fx)
+
+! ==== Description
+!! Impure wrapper function for `f_dst_llogistic_pdf_core`.
+!! Handles optional arguments and invalid values for arguments.
+
+! ==== Declarations
+  real(wp), intent(in)           :: x       !! sample position
+  real(wp), intent(in), optional :: mu      !! distribution mean (log-scale location)
+  real(wp), intent(in), optional :: alpha   !! distribution scale
+  real(wp), intent(in), optional :: beta    !! distribution shape
+  real(wp)                       :: mu_w    !! final value for mu
+  real(wp)                       :: alpha_w !! final value for alpha
+  real(wp)                       :: beta_w  !! final value for beta
+  real(wp)                       :: fx
+
+! ==== Instructions
+
+! ---- handle input
+
[…]
