### 2026-01-18 3ae5c64
(log) logistic docs and api update
