+++ src/fsml_dst.f90
@@ -2513,7 +2513,7 @@ end function f_dst_logistic_ppf_core
 
 ! ==================================================================== !
 ! -------------------------------------------------------------------- !
-impure function f_dst_llogistic_pdf(x, mu, alpha, beta) result(fx)
+impure function f_dst_llogistic_pdf(x, alpha, beta, loc) result(fx)
 
 ! ==== Description
 !! Impure wrapper function for `f_dst_llogistic_pdf_core`.
@@ -2521,10 +2521,10 @@ impure function f_dst_llogistic_pdf(x, mu, alpha, beta) result(fx)
 
 ! ==== Declarations
   real(wp), intent(in)           :: x       !! sample position
-  real(wp), intent(in), optional :: mu      !! distribution mean (log-scale location)
   real(wp), intent(in), optional :: alpha   !! distribution scale
   real(wp), intent(in), optional :: beta    !! distribution shape
-  real(wp)                       :: mu_w    !! final value for mu
+  real(wp), intent(in), optional :: loc     !! distribution location
+  real(wp)                       :: loc_w    !! final value for loc
   real(wp)                       :: alpha_w !! final value for alpha
   real(wp)                       :: beta_w  !! final value for beta
   real(wp)                       :: fx
@@ -2534,8 +2534,8 @@ impure function f_dst_llogistic_pdf(x, mu, alpha, beta) result(fx)
 ! ---- handle input
 
   ! assume mean/location = 0 (log-alpha), overwrite if specified
-  mu_w = 0.0_wp
-  if (present(mu)) mu_w = mu
+  loc_w = 0.0_wp
+  if (present(mu)) loc_w = loc
 
   ! assume alpha = 1, overwrite if specified
   alpha_w = 1.0_wp
@@ -2561,23 +2561,23 @@ impure function f_dst_llogistic_pdf(x, mu, alpha, beta) result(fx)
 ! ---- compute PDF
 
   ! call pure function to calculate probability/fx
-  fx = f_dst_llogistic_pdf_core(x, mu_w, alpha_w, beta_w)
+  fx = f_dst_llogistic_pdf_core(x, alpha_w, beta_w, loc_w)
 
 end function f_dst_llogistic_pdf
 
 
 ! ==================================================================== !
 ! -------------------------------------------------------------------- !
-elemental function f_dst_llogistic_pdf_core(x, mu, alpha, beta) result(fx)
+elemental function f_dst_llogistic_pdf_core(x, alpha, beta, loc) result(fx)
 
 ! ==== Description
 !! Probability density function for log-logistic distribution.
 
 ! ==== Declarations
   real(wp), intent(in) :: x     !! sample position
-  real(wp), intent(in) :: mu    !! distribution mean (log-alpha location)
   real(wp), intent(in) :: alpha !! distribution scale
   real(wp), intent(in) :: beta  !! distribution shape
+  real(wp), intent(in) :: loc   !! distribution location
   real(wp)             :: z     !! z-score (log-alpha)
   real(wp)             :: t     !! transformed variable
   real(wp)             :: fx
@@ -2585,7 +2585,7 @@ elemental function f_dst_llogistic_pdf_core(x, mu, alpha, beta) result(fx)
 ! ==== Instructions
 
   ! compute transformed variable
-  z = (log(x) - mu) / alpha
+  z = (log(x) - loc) / alpha
   t = exp(beta * z)
 
   ! calculate probability/fx
@@ -2596,7 +2596,7 @@ end function f_dst_llogistic_pdf_core
 
 ! ==================================================================== !
 ! -------------------------------------------------------------------- !
[…]
