+++ doc/api/dst.md
@@ -33,7 +33,7 @@ The location parameter \( \mu \) (`mu`) and scale parameter \( \sigma \) (`sigma
 
 `sigma`: An optional argument and positive scalar of type `real`. If passed, it must be non-zero positive. It will default to *1.0* if not passed.
 
-Invalid argument values will result in the return of a sentinel value.
+Invalid argument values will result in the return of a sentinel value (NaN).
 
 ### Returns
 The result is a scalar of the same type as `x`.
@@ -60,7 +60,7 @@ The location parameter \( \mu \) (`mu`), scale parameter \( \sigma \) (`sigma`),
 
 `tail`: An optional argument and `character` string. If passed, it must be one of the following: *"left"*, *"right"*, *"two"*, or *"confidence"*. If not passed, it will default to *"left"*.
 
-Invalid argument values will result in the return of a sentinel value.
+Invalid argument values will result in the return of a sentinel value (NaN).
 
 ### Returns
 The result is a scalar of the same type as `x`.
@@ -89,7 +89,7 @@ will return large negative or positive numbers (highly dependent on the toleranc
 
 `sigma`: An optional argument and positive scalar of type `real`. If passed, it must be non-zero positive. It will default to *1.0* if not passed.
 
-Invalid argument values will result in the return of a sentinel value.
+Invalid argument values will result in the return of a sentinel value (NaN).
 
 ### Returns
 The result is a scalar of the same type as `p`.
@@ -122,7 +122,7 @@ The location parameter \( \mu \) (`mu`) and scale parameter \( \sigma \) (`sigma
 
 `sigma`: An optional argument and positive scalar of type `real`. If passed, it must be non-zero positive. It will default to *1.0* if not passed.
 
-Invalid argument values will result in the return of a sentinel value.
+Invalid argument values will result in the return of a sentinel value (NaN).
 
 ### Returns
 The result is a scalar of the same type as `x`.
@@ -150,7 +150,7 @@ The location parameter \( \mu \) (`mu`), scale parameter \( \sigma \) (`sigma`),
 
 `tail`: An optional argument and `character` string. If passed, it must be one of the following: *"left"*, *"right"*, *"two"*, or *"confidence"*. If not passed, it will default to *"left"*.
 
-Invalid argument values will result in the return of a sentinel value.
+Invalid argument values will result in the return of a sentinel value (NaN).
 
 ### Returns
 The result is a scalar of the same type as `x`.
@@ -181,7 +181,7 @@ will return large negative or positive numbers (highly dependent on the toleranc
 
 `sigma`: An optional argument and positive scalar of type `real`. If passed, it must be non-zero positive. It will default to *1.0* if not passed.
 
-Invalid argument values will result in the return of a sentinel value.
+Invalid argument values will result in the return of a sentinel value (NaN).
 
 ### Returns
 The result is a scalar of the same type as `p`.
@@ -218,7 +218,7 @@ The scale parameters \( \alpha \) (`alpha`) and \( \beta \) (`beta`) and the loc
 
 `loc`:  A scalar of type `real`.
 
-Invalid argument values will result in the return of a sentinel value.
+Invalid argument values will result in the return of a sentinel value (NaN).
 
 ### Returns
 The result is a scalar of the same type as `x`.
@@ -246,7 +246,7 @@ The scale parameters \( \alpha \) (`alpha`) and \( \beta \) (`beta`), the locati
 
 `tail`: An optional argument and `character` string. If passed, it must be one of the following: *"left"*, *"right"*, *"two"*, or *"confidence"*. If not passed, it will default to *"left"*.
 
[…]
