### 2026-01-18 a287f34
(log) logistic examples and corrected comments
