### 2026-01-18 080bf3b
(log) logistic tests
