+++ test/test_tst.f90
@@ -54,9 +54,9 @@ subroutine handle_status(status)
 
 ! Instructions
   if (status) then
-     print*, "  passed"
+     print*, "  [✓] passed"
   else
-     print*, "  [error] one or more failed"
+     print*, "  [x] one or more failed"
      stop
   endif
 
