+++ test/test_dst.f90
@@ -52,6 +52,14 @@ program test_dst
   status = test_gpd(tol)
   call handle_status(status)
 
+  print*, "> dst: testing all logistic distribution functions (pdf, cdf, ppf)"
+  status = test_logistic(tol)
+  call handle_status(status)
+
+  print*, "> dst: testing all log-logistic distribution functions (pdf, cdf, ppf)"
+  status = test_llogistic(tol)
+  call handle_status(status)
+
   print*, "> dst: testing results of invalid arguments (and print messages)"
   status = test_invalid()
   call handle_status(status)
@@ -71,9 +79,9 @@ subroutine handle_status(status)
 
 ! Instructions
   if (status) then
-     print*, "  passed"
+     print*, "  [✓] passed"
   else
-     print*, "  [error] one or more failed"
+     print*, "  [x] one or more failed"
      stop
   endif
 
@@ -213,17 +221,17 @@ function test_t(tol) result(status)
      j = j + 1
      res = fsml_t_pdf(x(i), df(i))
      if (abs( res - ans(j) ) .gt. tol) status = .false.
-  end do
+  enddo
   do i = 1, n
      j = j + 1
      res = fsml_t_cdf(x(i), df(i))
      if (abs( res - ans(j) ) .gt. tol) status = .false.
-  end do
+  enddo
   do i = 1, n
      j = j + 1
      res = fsml_t_ppf(p(i), df(i))
      if (abs( res - ans(j) ) .gt. tol) status = .false.
-  end do
+  enddo
 
 end function test_t
 
@@ -290,17 +298,17 @@ function test_gamma(tol) result(status)
      j = j + 1
      res = fsml_gamma_pdf(x(i), alpha(i), beta(i))
      if (abs( res - ans(j) ) .gt. tol) status = .false.
-  end do
+  enddo
   do i = 1, n
      j = j + 1
      res = fsml_gamma_cdf(x(i), alpha(i), beta(i))
      if (abs( res - ans(j) ) .gt. tol) status = .false.
-  end do
+  enddo
   do i = 1, n
      j = j + 1
      res = fsml_gamma_ppf(p(i), alpha(i), beta(i))
      if (abs( res - ans(j) ) .gt. tol) status = .false.
-  end do
+  enddo
 
 end function test_gamma
 
@@ -361,17 +369,17 @@ function test_exp(tol) result(status)
      j = j + 1
      res = fsml_exp_pdf(x(i), lambda(i))
      if (abs( res - ans(j) ) .gt. tol) status = .false.
-  end do
+  enddo
   do i = 1, n
      j = j + 1
      res = fsml_exp_cdf(x(i), lambda(i))
      if (abs( res - ans(j) ) .gt. tol) status = .false.
-  end do
+  enddo
   do i = 1, n
      j = j + 1
      res = fsml_exp_ppf(p(i), lambda(i))
      if (abs( res - ans(j) ) .gt. tol) status = .false.
-  end do
+  enddo
 
 end function test_exp
 
@@ -432,17 +440,17 @@ function test_chi2(tol) result(status)
      j = j + 1
      res = fsml_chi2_pdf(x(i), df(i))
      if (abs( res - ans(j) ) .gt. tol) status = .false.
-  end do
+  enddo
   do i = 1, n
      j = j + 1
      res = fsml_chi2_cdf(x(i), df(i))
      if (abs( res - ans(j) ) .gt. tol) status = .false.
-  end do
+  enddo
   do i = 1, n
      j = j + 1
      res = fsml_chi2_ppf(p(i), df(i))
      if (abs( res - ans(j) ) .gt. tol) status = .false.
-  end do
+  enddo
 
[…]
