### 2026-02-03 f578d27
fixed gamma loc handling
