+++ releases/draft_v0.1.1/notes.md
@@ -4,8 +4,7 @@
 The following has changed since the last release (v0.1.0):
 
 ### Bug Fixes
-- Fixed gamma distribution loc bug and removed redundant arg checks in pure procedure ([f578d27](https://github.com/sebastian-mutz/fsml/commit/f578d27fba2518bb9b0f66056f2a59fb233fbf25))
-
+- Fixed gamma distribution loc bug and removed redundant arg checks in pure procedure ([f578d27](https://github.com/sebastian-mutz/fsml/commit/f578d27fba2518bb9b0f66056f2a59fb233fbf25)) [contributed by @sebastian-mutz]
 
 ### New Features
 
@@ -14,20 +13,20 @@ The following has changed since the last release (v0.1.0):
 - [ ] Added Canonical Correlation Analysis (CCA) [contributed by @loiseaujc].
 
 #### DST module
-- [x] Added logistic distribution PDF, CDF, and PPF. ([3f81edb](https://github.com/sebastian-mutz/fsml/commit/3f81edb9e9ecfe48a34a2d4a75115fb909d9cccc), [3ae5c64](https://github.com/sebastian-mutz/fsml/commit/3ae5c6413967e13ff11495414b4d83f26f103453), [a287f34](https://github.com/sebastian-mutz/fsml/commit/a287f34ca90b05c863bb840351e641e5f36cab02), [080bf3b](https://github.com/sebastian-mutz/fsml/commit/080bf3b2f8cc6da3158bcdc05418d3c9792835c7))
-- [x] Added log-logistic distribution PDF, CDF, and PPF. ([3ad09b9](https://github.com/sebastian-mutz/fsml/commit/3ad09b9107c1aaa6f5d0516491d32a1ac3d7999), [3ae5c64](https://github.com/sebastian-mutz/fsml/commit/3ae5c6413967e13ff11495414b4d83f26f103453), [a287f34](https://github.com/sebastian-mutz/fsml/commit/a287f34ca90b05c863bb840351e641e5f36cab02), [080bf3b](https://github.com/sebastian-mutz/fsml/commit/080bf3b2f8cc6da3158bcdc05418d3c9792835c7))
+- [x] Added logistic distribution PDF, CDF, and PPF. ([3f81edb](https://github.com/sebastian-mutz/fsml/commit/3f81edb9e9ecfe48a34a2d4a75115fb909d9cccc), [3ae5c64](https://github.com/sebastian-mutz/fsml/commit/3ae5c6413967e13ff11495414b4d83f26f103453), [a287f34](https://github.com/sebastian-mutz/fsml/commit/a287f34ca90b05c863bb840351e641e5f36cab02), [080bf3b](https://github.com/sebastian-mutz/fsml/commit/080bf3b2f8cc6da3158bcdc05418d3c9792835c7)) [contributed by @sebastian-mutz]
+- [x] Added log-logistic distribution PDF, CDF, and PPF. ([3ad09b9](https://github.com/sebastian-mutz/fsml/commit/3ad09b9107c1aaa6f5d0516491d32a1ac3d7999), [3ae5c64](https://github.com/sebastian-mutz/fsml/commit/3ae5c6413967e13ff11495414b4d83f26f103453), [a287f34](https://github.com/sebastian-mutz/fsml/commit/a287f34ca90b05c863bb840351e641e5f36cab02), [080bf3b](https://github.com/sebastian-mutz/fsml/commit/080bf3b2f8cc6da3158bcdc05418d3c9792835c7)) [contributed by @sebastian-mutz]
 
 #### NLP module
 
-- [ ] Implemented Lance-Williams algorithm for clustering; more efficient than re-computing Mahalanobis distance
[…]
